Updates the code to always create variables and computations of the same dtype as the its inputs. Previously, if float64 was enabled, some of the results would be (potentially incorrectly) promoted to higher precision.
PiperOrigin-RevId: 508391517

+++ kfac_jax/_src/utils/math.py
@@ -324,7 +324,7 @@ def psd_inv_cholesky(matrix: chex.Array, damping: chex.Array) -> chex.Array:
   if matrix.shape[:1] != matrix.shape[1:]:
     raise ValueError(f"Expected square matrix, but got shape {matrix.shape}.")
 
-  identity = jnp.eye(matrix.shape[0])
+  identity = jnp.eye(matrix.shape[0], dtype=matrix.dtype)
 
   return linalg.solve(matrix + damping * identity, identity, assume_a="pos")
 
@@ -377,6 +377,7 @@ def pi_adjusted_kronecker_inverse(
 
   # kron(arrays) = c * kron(us)
   c = jnp.exp(jnp.sum(jnp.log(jnp.stack(norms)) - jnp.log(jnp.stack(dims))))
+  damping = damping.astype(c.dtype)
 
   def regular_inverse() -> Tuple[chex.Array, ...]:
 
@@ -417,7 +418,7 @@ def pi_adjusted_kronecker_inverse(
     for a in us:
 
       if a.ndim == 2:
-        inv = jnp.eye(a.shape[0])
+        inv = jnp.eye(a.shape[0], dtype=a.dtype)
 
       else:
         inv = jnp.ones_like(a)
@@ -643,7 +644,8 @@ def safe_psd_eigh(
   # of cuda and cudablas they can cause a runtime error.
   s, q = lax.cond(
       jnp.any(jnp.isnan(x)),
-      lambda _: (jnp.full([d], jnp.nan), jnp.full([d, d], jnp.nan)),
+      lambda _: (jnp.full([d], jnp.nan, dtype=x.dtype),  # pylint: disable=g-long-lambda
+                 jnp.full([d, d], jnp.nan, dtype=x.dtype)),
       functools.partial(_eigh, force_on_host=force_on_host),
       x,
   )
